Abstract collection params
